# src/modules/coreHttpApi/openapi.yml
openapi: 3.0.0
servers:
  - url: /
    description: Local Connector API
info:
  version: "0.4"
  title: Connector
  description: |
    This API is work in progress and might be changed without further notice on any time.

    You can retrieve the [yaml](/docs/yaml) or [json](/docs/json) representation of this specification by following the corresponding links.

    You can also view this documenation in [rapidoc](/docs/rapidoc).

  license:
    name: MIT
    url: https://raw.githubusercontent.com/nmshd/cns-connector/main/LICENSE
  contact:
    name: j&s-soft GmbH
    email: info@js-soft.com
    url: https://www.js-soft.com/
paths:
...
  /Monitoring/Requests:
    get:
      operationId: getRequests
      description: Show the number of requests and the status codes that were returned by the Connector.
      tags:
        - Monitoring
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  since:
                    type: string
                    format: date-time
                    description: The time since when the data was collected (typically this is the startup time of the Connector).
                  requestCount:
                    type: number
                    example: 507
                    description: The number of requests the Connector has processed.
                  requestCountByStatus:
                    type: object
                    additionalProperties:
                      type: number
                    example: { "200": 500, "401": 2, "500": 5 }
                required:
                  - since
                  - requestCount
                  - responseCountByStatus
          headers:
            X-Response-Duration-ms:
              schema:
                $ref: "#/components/schemas/HeaderContent_X-Response-Duration-ms"
            X-Response-Time:
              schema:
                $ref: "#/components/schemas/HeaderContent_X-Response-Time"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
  /Monitoring/Support:
    get:
      operationId: getSupportInfo
      description: Show support Information.
      tags:
        - Monitoring
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  version:
                    type: object
                    $ref: "#/components/schemas/ConnectorVersion"
                  health:
                    type: object
                    $ref: "#/components/schemas/ConnectorHealth"
                  configuration:
                    type: object
                    example:
                      {
                        "database": { "connectionString": "**********************************", "dbName": "accountName" },
                        "modules": { "httpServer": { "enabled": true }, "coreHttpApi": { "enabled": true } }
                      }
                  identityInfo:
                    type: object
                    $ref: "#/components/schemas/IdentityInfo"
                required:
                  - result
          headers:
            X-Response-Duration-ms:
              schema:
                $ref: "#/components/schemas/HeaderContent_X-Response-Duration-ms"
            X-Response-Time:
              schema:
                $ref: "#/components/schemas/HeaderContent_X-Response-Time"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"

  # ------------------- Account -------------------
...
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
  schemas:
...
    Address:
      type: string
      format: Address
      minLength: 35
      maxLength: 36
      example: id_________________________________
...
    IdentityInfo:
      type: object
      properties:
        address:
          allOf:
            - $ref: "#/components/schemas/Address"
          description: The enmeshed address of the Connector.
          nullable: false
        publicKey:
          type: string
          description: The public key of the Connector used to digitally sign data.
          nullable: false
      required:
        - address
        - publicKey
    ConnectorHealth:
      type: object
      properties:
        isHealthy:
          type: boolean
          nullable: false
        services:
          type: object
          nullable: false
          properties:
            database:
              type: string
              enum: [healthy, unhealthy]
            backbone:
              type: string
              enum: [healthy, unhealthy]
          required:
            - database
            - backbone
      required:
        - isHealthy
        - services
    ConnectorVersion:
      type: object
      properties:
        build:
          type: string
          example: 50
          description: The build number of the Connector Docker image.
        commit:
          type: string
          example: c459aa1b74a6b51d1230869ec8b199d4c7506679
          description: The commit that triggered the build of the Connector docker-image.
        version:
          type: string
          example: 1.0.0
          description: The Connector version.
        date:
          type: string
          format: date-time
          description: The time the Connector docker-image was built.
      required:
        - build
        - commit
        - version
        - date
...
    ErrorContent:
      type: object
      properties:
        error:
          type: object
          properties:
            id:
              description: A unique ID for the error instance.
              type: string
            code:
              description: >
                The error code.  

                There are some generic error codes like: 
                  - error.connector.recordNotFound
                  - error.connector.unauthorized
                  - error.connector.forbidden
                  - error.connector.validation.missingProperty
                  - error.connector.validation.invalidPropertyValueLength
                  - error.connector.validation.invalidPropertyValue

                Further there are route specific error codes.
              type: string
            message:
              description: An error message having more details about what went wrong. This message is not localized.
              type: string
            docs:
              description: A link to additional documentation about the error.
              type: string
            time:
              description: A timestamp that describes when the error occured.
              type: string
          required:
            - id
            - code
            - message
            - docs
            - time
      required:
        - error
    HeaderContent_X-Response-Duration-ms:
      description: The time in milliseconds the request took to be processed by the platform.
      type: number
    HeaderContent_X-Response-Time:
      description: A timestamp that describes when the response was sent to the client.
      type: string
      format: number
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorContent"
      headers:
        X-Response-Duration-ms:
          schema:
            $ref: "#/components/schemas/HeaderContent_X-Response-Duration-ms"
        X-Response-Time:
          schema:
            $ref: "#/components/schemas/HeaderContent_X-Response-Time"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorContent"
      headers:
        X-Response-Duration-ms:
          schema:
            $ref: "#/components/schemas/HeaderContent_X-Response-Duration-ms"
        X-Response-Time:
          schema:
            $ref: "#/components/schemas/HeaderContent_X-Response-Time"
...